+++ cabal-to-dhall/Main.hs
@@ -4,21 +4,15 @@
 module Main ( main ) where
 
 import Control.Applicative ( (<**>), optional )
-import Data.Version ( showVersion )
-import GHC.Stack
 
 import qualified Data.ByteString as ByteString
-import qualified Data.Text.Lazy as LazyText
-import qualified Data.Text.Lazy.IO as LazyText
 import qualified Data.Text.Prettyprint.Doc as Pretty
 import qualified Data.Text.Prettyprint.Doc.Render.Text as Pretty
-import qualified Data.Text.Prettyprint.Doc.Symbols.Unicode as Pretty
-import qualified Dhall.Core
 import qualified Options.Applicative as OptParse
 import qualified System.IO
 
-import CabalToDhall ( cabalToDhall, DhallLocation ( DhallLocation ) )
-import qualified Paths_dhall_to_cabal as Paths
+import CabalToDhall ( cabalToDhall )
+import DhallLocation ( dhallFromGitHub )
 
 
 data Command
@@ -60,58 +54,8 @@ main = do
       runCabalToDhall options
 
 
-version :: LazyText.Text
-version = LazyText.pack ( showVersion Paths.version )
-
-
-preludeLocation :: Dhall.Core.Import
-preludeLocation =
-  Dhall.Core.Import
-    { Dhall.Core.importHashed =
-        Dhall.Core.ImportHashed
-          { Dhall.Core.hash =
-              Nothing
-          , Dhall.Core.importType =
-              Dhall.Core.URL
-                "https://raw.githubusercontent.com"
-                ( Dhall.Core.File
-                   ( Dhall.Core.Directory [ "dhall", version, "dhall-to-cabal", "dhall-lang" ] )
-                   "prelude.dhall"
-                )
-                ""
-                Nothing
-          }
-    , Dhall.Core.importMode =
-        Dhall.Core.Code
-    }
-
-
-typesLocation :: Dhall.Core.Import
-typesLocation =
-  Dhall.Core.Import
-    { Dhall.Core.importHashed =
-        Dhall.Core.ImportHashed
-          { Dhall.Core.hash =
-              Nothing
-          , Dhall.Core.importType =
-              Dhall.Core.URL
-                "https://raw.githubusercontent.com"
-                ( Dhall.Core.File
-                   ( Dhall.Core.Directory [ "dhall", version, "dhall-to-cabal", "dhall-lang" ] )
-                   "types.dhall"
-                )
-                ""
-                Nothing
-          }
-    , Dhall.Core.importMode =
-        Dhall.Core.Code
-    }
-
-
 runCabalToDhall :: CabalToDhallOptions -> IO ()
 runCabalToDhall CabalToDhallOptions{ cabalFilePath } = do
-  let dhallLocation = DhallLocation preludeLocation typesLocation
-
   source <-
     case cabalFilePath of
       Nothing ->
@@ -121,7 +65,7 @@ runCabalToDhall CabalToDhallOptions{ cabalFilePath } = do
         ByteString.readFile filePath
 
   dhall <-
-    cabalToDhall dhallLocation source
+    cabalToDhall dhallFromGitHub source
 
   Pretty.renderIO
     System.IO.stdout